+++ src/server/src/server-glue/global-ext.d.ts
@@ -1,0 +1,9 @@
+declare var nodeRequire: NodeRequire
+declare namespace NodeJS {
+  export interface Process {
+    browser: boolean
+  }
+  export interface Global {
+    nodeRequire: NodeRequire
+  }
+}